modular: seperated skill data in Skills component
